Move format_string_with_location to public API
